added hangman
I opted to not purchase the JavaScript for Kids tutorial so instead I completed the Hangman course on Code Academy. This code was taken from Code Academy but I have added comments to demonstrate my understanding of the logic and through the course have ultimately typed out all of the code with the notable exception of the CSS which I copied from the Code Academy course. If this is unsatisfactory in anyway please let me know.